backend/requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0

# AI/ML Libraries
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Document Processing
unstructured==0.12.0
pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10

# Language Detection
langdetect==1.0.9

# Online AI API
openai==1.10.0
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6

# Testing
requests==2.31.0
colorama==0.4.6

# Optional: For better PDF processing
# poppler-utils (system package, install via apt/brew/choco)
# tesseract-ocr (system package, install via apt/brew/choco)